Refactor on mac. Passing mac tests.
